+++ src/test/scala/objects/AutoRepairTest.scala
@@ -142,11 +142,10 @@ class AutoRepairRequestNtuRepeatTest extends FreedContextActorTest {
       (0 to 3).foreach { _ =>
         val buildingMsg = buildingProbe.receiveOne(max = 1000 milliseconds)
         assert(buildingMsg match {
-          case BuildingActor.Ntu(NtuCommand.Request(drain, _)) =>
-            drain == terminal.Definition.autoRepair.get.drain
-          case _ =>
-            false
+          case BuildingActor.Ntu(NtuCommand.Request(_, _)) => true
+          case _                                           => false
         })
+        terminal.Actor ! NtuCommand.Grant(null, 0)
       }
     }
   }
